Host clock setting hack for a specific project.
Dynamic clock management on RPi leads the ARM core to be clocked from 600MHz up to 1400MHz
(on a RPi 3B+). This causes large deviations of the JTAG/SWD clock and makes the JTAG/SWD
link unreliable. The target is no longer able to cope with clock pulses.

This hack freezes the clock of all the RPi CPU cores to the lowest value (600MHz on RPi 3B+)
so that the JTAG/SWD clock is known and predictable. It also attempts to restore the initial
CPU clock scaling governor on exit.

This is an ugly hack, but bit-banging GPIO @ several MHZ from user space as openocd does is
already a large one :-)

As long as openocd is running, the CPU computational power is reduced when RPi GPIO is
selected. This is not a big deal for me - YMMV :-)